requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
lxml==5.3.2
playwright==1.51.0
pydub==0.25.1
pyee==12.1.1
requests==2.32.3
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
google-cloud-texttospeech
